Restructuration des pages de badges
