+++ locales/en.json
@@ -174,6 +174,8 @@
     "userNotFound": "Nobody matches that name.",
     "badgeNotFound": "No badge found with that name.",
     "searchNotFound": "Your search for {search} found no results.",
-    "private": "Private"
+    "private": "Private",
+    "errorPing": "OpenBadgeFactory services are currently unavailable, so it is not possible to issue badges!",
+    "list": "Badges list"
   }
 }
