+++ locales/fr.json
@@ -174,6 +174,8 @@
     "userNotFound": "Personne ne correspond à ce nom.",
     "badgeNotFound": "Aucun badge trouvé avec ce nom.",
     "searchNotFound": "Votre recherche pour {search} ne donne aucun résultat.",
-    "private": "Privé"
+    "private": "Privé",
+    "pingError": "Les services OpenBadgeFactory sont actuellement indisponible, il n'est donc pas possible d'émettre des badges !",
+    "list": "Liste des badges"
   }
 }
